Feat: Basic Missile & Tower Script
In this commit, there should be two C# scripts
 - Missile.cs
 - TowerController.cs

The missile script makes a missile move towards a end point (target) that is set by the tower controller. The tower controller checks if a mouse click happens and the coordinates where it happened. Then it will instantiate a missile prefab, and thus the missile will move to its target.

To test this, you can make a 3D cube game object and give it the tower controller, and within the cube make sure you also give the missile prefab to its script component. Then you hit play and click on the screen, and a missile should fire from the tower.