Removed Ghost Fire Particles
Removed Ghost Fire Particles and added Animation to make when the ghost is firing a missile it will Scale slight on the Y axis and the X axis. Also changed the material of the ghost when firing to red to be an indicator to the player that the ghost is firing a missile.